Adjusted GenericEntity.java and Map.java.
Added an overridable method in GenericEntity.java - canEntityMoveOnto(GenericEntity) so that the default for an entity being able to move is false.  The function should be overridden in extended classes to account for how different entities should interact.

Map.java now has some extra notes about how stuff works.  The array order is now different for optimization purposes ([mapHeight][mapWidth] instead of [mapWidth][mapHeight]) and the block divisor is given as a final variable rather than randomly placed throughout the code.  Also, a method for finding what is in a particular position was added, as was a method to determine if a position is empty or not.