Started a method in Map to move entities.
It has a bunch of commends in it to show what needs to be done.  I suspect that it might be better to do it in a slightly different way though, where the movement isn't split into 3 or something.  Perhaps the surrounding blocks should just refer to the centre position block, or only centre position blocks should be used and everything should check the 8 surrounding blocks.  Probably the former of those options.

Changed a comment in Position.java to indicate that I think the emptySpace() method should be moved to the map class.